Partial file separation
